💄 style(agent-chat): move the terminal button into the header right actions
The terminal toggle sat next to the topic title on the left, away from
every other header action. Extract it into its own TerminalPanelToggle
component and render it in the right slot beside the IDE / share /
panel buttons, using the shared DESKTOP_HEADER_ICON_SMALL_SIZE so it
lines up with its neighbours.
