+++ src/routes/(main)/agent/features/Conversation/Header/HeaderActions/index.test.tsx
@@ -1,24 +1,8 @@
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import type { ReactNode } from 'react';
 import { describe, expect, it, vi } from 'vitest';
 
 import HeaderActions from './index';
-
-const { toggleTerminalPanel } = vi.hoisted(() => ({
-  toggleTerminalPanel: vi.fn(),
-}));
-
-vi.mock('@/const/version', () => ({ isDesktop: true }));
-
-vi.mock('@/store/global', () => ({
-  useGlobalStore: (
-    selector: (state: { toggleTerminalPanel: typeof toggleTerminalPanel }) => unknown,
-  ) => selector({ toggleTerminalPanel }),
-}));
-
-vi.mock('react-i18next', () => ({
-  useTranslation: () => ({ t: (key: string) => key }),
-}));
 
 vi.mock('@lobehub/ui', () => ({
   ActionIcon: ({ title, onClick }: { title?: string; onClick?: () => void }) => (
@@ -54,13 +38,5 @@ describe('Conversation header actions', () => {
     render(<HeaderActions />);
 
     expect(screen.getByTestId('topic-info-header')).toBeInTheDocument();
-  });
-
-  it('opens the terminal directly from the desktop header', () => {
-    render(<HeaderActions />);
-
-    fireEvent.click(screen.getByRole('button', { name: 'terminalPanel.title' }));
-
-    expect(toggleTerminalPanel).toHaveBeenCalledWith(true);
   });
 });
