+++ src/routes/(main)/agent/features/Conversation/Header/index.tsx
@@ -17,6 +17,7 @@ import { useElectronStore } from '@/store/electron';
 import HeaderActions from './HeaderActions';
 import ShareButton from './ShareButton';
 import Tags from './Tags';
+import TerminalPanelToggle from './TerminalPanelToggle';
 import WorkingPanelToggle from './WorkingPanelToggle';
 
 // Below this column width the header is a solid in-flow bar with a bottom
@@ -181,6 +182,7 @@ const Header = memo(() => {
             {isLocalSystemEnabled && workingDirectory && (
               <OpenInAppButton workingDirectory={workingDirectory} />
             )}
+            <TerminalPanelToggle />
             <TopicCommentButton />
             <ShareButton />
             <WorkingPanelToggle />
